fix(webhook): correct status
