+++ packages/backend/src/app/webhooks/webhook-controller.ts
@@ -1,8 +1,10 @@
 import { FastifyPluginAsync, FastifyRequest } from 'fastify'
 import { StatusCodes } from 'http-status-codes'
-import { WebhookUrlParams } from '@activepieces/shared'
+import { ActivepiecesError, ErrorCode, Flow, FlowId, WebhookUrlParams } from '@activepieces/shared'
 import { webhookService } from './webhook-service'
 import { captureException, logger } from '../helper/logger'
+import { isNil } from 'lodash'
+import { flowRepo } from '../flows/flow/flow.repo'
 
 export const webhookController: FastifyPluginAsync = async (app) => {
     app.all(
@@ -13,7 +15,8 @@ export const webhookController: FastifyPluginAsync = async (app) => {
             },
         },
         async (request: FastifyRequest<{ Params: WebhookUrlParams }>, reply) => {
-            handler(request, request.params.flowId)
+            const flow = await getFlowOrThrow(request.params.flowId)
+            handler(request, flow)
             await reply.status(StatusCodes.OK).send()
         },
     )
@@ -26,7 +29,8 @@ export const webhookController: FastifyPluginAsync = async (app) => {
             },
         },
         async (request: FastifyRequest<{ Querystring: WebhookUrlParams }>, reply) => {
-            handler(request, request.query.flowId)
+            const flow = await getFlowOrThrow(request.query.flowId)
+            handler(request, flow)
             await reply.status(StatusCodes.OK).send()
         },
     )
@@ -40,9 +44,9 @@ export const webhookController: FastifyPluginAsync = async (app) => {
         },
         async (request: FastifyRequest<{ Params: WebhookUrlParams }>, reply) => {
             logger.debug(`[WebhookController#simulate] flowId=${request.params.flowId}`)
-
+            const flow = await getFlowOrThrow(request.params.flowId)
             await webhookService.simulationCallback({
-                flowId: request.params.flowId,
+                flow: flow,
                 payload: {
                     method: request.method,
                     headers: request.headers as Record<string, string>,
@@ -56,11 +60,11 @@ export const webhookController: FastifyPluginAsync = async (app) => {
     )
 }
 
-const handler = async (request: FastifyRequest, flowId: string) => {
+const handler = async (request: FastifyRequest, flow: Flow) => {
     // If we don't catch the error here, it will crash the Fastify API. Adding await before the function call can help, but since 3P services expect a fast response, we still don't want to wait for the callback to finish.
     try {
         await webhookService.callback({
-            flowId: flowId,
+            flow: flow,
             payload: {
                 method: request.method,
                 headers: request.headers as Record<string, string>,
@@ -72,4 +76,31 @@ const handler = async (request: FastifyRequest, flowId: string) => {
     catch (e) {
         captureException(e)
     }
+}
+
+const getFlowOrThrow = async (flowId: FlowId): Promise<Flow> => {
+    if (isNil(flowId)) {
+        logger.error('[WebhookService#getFlowOrThrow] error=flow_id_is_undefined')
+        throw new ActivepiecesError({
+            code: ErrorCode.VALIDATION,
+            params: {
+                message: 'flowId is undefined',
+            },
+        })
+    }
+
+    const flow = await flowRepo.findOneBy({ id: flowId })
+
+    if (isNil(flow)) {
+        logger.error(`[WebhookService#getFlowOrThrow] error=flow_not_found flowId=${flowId}`)
+
+        throw new ActivepiecesError({
+            code: ErrorCode.FLOW_NOT_FOUND,
+            params: {
+                id: flowId,
+            },
+        })
+    }
+
+    return flow
 }
